move `PeriodicTableDemo` to `src/site` and import on landing and /periodic-table pages
