+++ src/site/index.ts
@@ -3,6 +3,7 @@ import type { IdStructure } from '../lib'
 export { default as DemoNav } from './DemoNav.svelte'
 export { default as Footer } from './Footer.svelte'
 export { default as PeriodicTableControls } from './PeriodicTableControls.svelte'
+export { default as PeriodicTableDemo } from './PeriodicTableDemo.svelte'
 
 export const structures = Object.entries(
   import.meta.glob(`./structures/*.json`, {
